API: move CachedVisionModel up to .vision
