+++ src/artkit/model/vision/base/__init__.py
@@ -18,5 +18,4 @@
 Base classes for the vision model
 """
 
-from ._adapter import *
 from ._vision import *
